Falta seleccionar pelicula correcta y añadir mas campos al mappeo

+++ src/main/java/ad/project/backend_wherewatch/domain/repositories/MovieRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface MovieRepository extends JpaRepository<Movie, Long>{
+public interface MovieRepository extends JpaRepository<Movie, Integer>{
     Optional<Movie> findByTitle(String title);
     List<Movie> findByTitleContainingIgnoreCase(String title);
 }
